Add configuration options for scheduled-feed binary
Feeds and publishers have been slightly reworked to support
configuration in place of constant endpoints. In addition to this
it is now possible to select a limited subset of feeds to poll.

Default behaviour has been preserved through a combination of default configurations
and mutating configurations post-load based on environment variables.