feeds/npm: Fix pkg versioning & corresponding creation date
The assumption that `pubDate` reported on the main rss feed is that
of the `latest` dist-tag is only always valid when there are no
other dist-tags for the package. The feed event can be triggered
by any of the listed dist-tags updating. Additional dist-tags can
be given arbitrary names, as such each tag has to be checked to
find which has most recent timestamp.

There is also the possibility that a given package produces multiple
feed events within a snapshot of the rss poll. The `unpublishing` of
packages that triggered feed events needs to be handled too.