feeds/: Allow for per package/feed error handling when polling
When errors occur but package data can still be processed then
the system should continue where possible. This is particularly
beneficial in feeds where data is fetched on a per package basis,
as a singular failure shouldn't immediately lead to data loss
for the whole session. `pollAndPublish()` error handling and
subsequent callers have been modified to handle this.

This also removes `errUnpublished` as a 'hard' error when polling
npm for critical packages as such it is only logged.